# index.js
import express from 'express';
import fs from 'fs';
import swaggerUi from 'swagger-ui-express';
const app = express();
const PORT = 3000;
app.use(express.json());

const swaggerDocument = JSON.parse(fs.readFileSync('./openapi.json', 'utf8'));
app.use('/docs', swaggerUi.serve, swaggerUi.setup(swaggerDocument));
let tasks = [
    {id: 1, title: "Learn Express.js", done: false},
    {id: 2, title: "Build a CRUD API", done: false},
    {id: 3, title: "Clean up .gitignore", done: false}
];
app.get('/', (req, res) => {
    res.json({
        name: "Task API",
        version: "1.0",
        endpoints: ["/tasks"]
    });
});
app.get('/health', (req, res) => {
    res.json({status: "ok"});
});
app.get('/tasks', (req, res) => {
    res.json(tasks);
});
app.get('/tasks/:id', (req, res) => {
    const taskId = parseInt(req.params.id);
    const task = tasks.find(t => t.id === taskId);
    if (!task) {
        return res.status(404).json({error: `Task ${req.params.id} not found`});
    }
    res.json(task);
});
app.post('/tasks', (req, res) => {
    const {title} = req.body;
    if (!title || typeof title !== 'string' || title.trim() === '') {
        return res.status(400).json({error: "Title is required and must be a non-empty string"});
    }
    const newId = tasks.length > 0 ? Math.max(...tasks.map(t => t.id)) + 1 : 1;
    const newTask = {
        id: newId, 
        title: title.trim(), 
        done: false
    };
    tasks.push(newTask);
    res.status(201).json(newTask);
});
app.put('/tasks/:id', (req, res) => {
    const taskId = parseInt(req.params.id, 10);
    const task = tasks.find(t => t.id === taskId);
    if (!task) {
        return res.status(404).json({ error: `Task ${req.params.id} not found` });
    }

    const { title, done } = req.body;

    if (title !== undefined && (typeof title !== 'string' || title.trim() === '')) {
        return res.status(400).json({ error: "Title must be a non-empty string" });
    }
    if (done !== undefined && typeof done !== 'boolean') {
        return res.status(400).json({ error: "Done must be a boolean" });
    }

    if (title !== undefined) task.title = title.trim();
    if (done !== undefined) task.done = done;

    res.json(task);
});

app.delete('/tasks/:id', (req, res) => {
    const taskId = parseInt(req.params.id, 10);
    const index = tasks.findIndex(t => t.id === taskId);
    if (index === -1) {
        return res.status(404).json({ error: `Task ${req.params.id} not found` });
    }
    tasks.splice(index, 1);
    res.status(204).send();
});
app.listen(PORT, () => {
    console.log(`Server is running on http://localhost:${PORT}`);
});